Update test details

+++ OsmAndMapCreator/src/net/osmand/data/diff/ObfDiffGenerator.java
@@ -7,6 +7,7 @@ import java.io.IOException;
 import java.sql.SQLException;
 import java.util.Collections;
 
+import net.osmand.binary.BinaryInspector;
 import net.osmand.binary.BinaryMapDataObject;
 import net.osmand.binary.BinaryMapIndexReader.MapIndex;
 import net.osmand.binary.BinaryMapRouteReaderAdapter.RouteRegion;
@@ -24,12 +25,13 @@ public class ObfDiffGenerator {
 	public static void main(String[] args) throws IOException, RTreeException {
 		if(args.length == 1 && args[0].equals("test")) {
 			args = new String[3];
-			args[0] = "/Users/victorshcherb/osmand/maps/diff/Ukraine_kiev-city_europe_09_2.obf";
-			args[1] = "/Users/victorshcherb/osmand/maps/diff/Ukraine_kiev-city_europe_09.obf";
+			args[0] = "/Users/victorshcherb/osmand/maps/diff/Belarus_europe_17_08_28O.obf";
+			args[1] = "/Users/victorshcherb/osmand/maps/diff/Belarus_europe_17_08_28N.obf";
 			args[2] = "/Users/victorshcherb/osmand/maps/diff/Diff.obf";
+//			args[2] = "stdout";
 		}
 		if (args.length != 3) {
-			System.out.println("Usage: <path to old obf> <path to new obf> <result file name>");
+			System.out.println("Usage: <path to old obf> <path to new obf> <[result file name] or [stdout]>");
 			System.exit(1);
 			return;
 		}
@@ -45,7 +47,7 @@ public class ObfDiffGenerator {
 	private void run(String[] args) throws IOException, RTreeException, SQLException {
 		File start = new File(args[0]);
 		File end = new File(args[1]);
-		File result  = new File(args[2]);
+		File result  = args.length < 3 || args[2].equals("stdout") ? null :new File(args[2]);
 		if (!start.exists() || !end.exists()) {
 			System.err.println("Input Obf file doesn't exist");
 			System.exit(1);
@@ -61,22 +63,24 @@ public class ObfDiffGenerator {
 		fEnd.readObfFiles(Collections.singletonList(end));
 
 		System.out.println("Comparing the files...");
-		compareMapData(fStart, fEnd);
+		compareMapData(fStart, fEnd, result == null);
 		// Compare Routing
-		compareRouteData(fStart, fEnd);
+		compareRouteData(fStart, fEnd, result == null);
 		// TODO Compare POI
 		// comparePOI(fStart, fEnd);
 		// TODO Compare Transport
 		//compareTransport(fStart, fEnd);
 		
 		System.out.println("Finished comparing.");
-		if (result.exists()) {
-			result.delete();
+		if (result != null) {
+			if (result.exists()) {
+				result.delete();
+			}
+			fEnd.writeFile(result);
 		}
-		fEnd.writeFile(result);
 	}
 
-	private void compareMapData(ObfFileInMemory fStart, ObfFileInMemory fEnd) {
+	private void compareMapData(ObfFileInMemory fStart, ObfFileInMemory fEnd, boolean print) {
 		fStart.filterAllZoomsBelow(13);
 		fEnd.filterAllZoomsBelow(13);
 		MapIndex mi = fEnd.getMapIndex();
@@ -91,29 +95,59 @@ public class ObfDiffGenerator {
 		for (MapZoomPair mz : fStart.getZooms()) {
 			TLongObjectHashMap<BinaryMapDataObject> startData = fStart.get(mz);
 			TLongObjectHashMap<BinaryMapDataObject> endData = fEnd.get(mz);
+			if(print) {
+				System.out.println("Compare map " + mz);
+			}
 			if (endData == null) {
 				continue;
 			}
 			for (Long idx : startData.keys()) {
 				BinaryMapDataObject objE = endData.get(idx);
 				BinaryMapDataObject objS = startData.get(idx);
-				if (objE == null) {
-					// Object with this id is not present in the second obf
-					BinaryMapDataObject obj = new BinaryMapDataObject(idx, objS.getCoordinates(), null,
-							objS.getObjectType(), objS.isArea(), new int[] { deleteId }, null);
-					endData.put(idx, obj);
-				} else if (objE.compareBinary(objS)) {
-					endData.remove(idx);
+				if (print) {
+					if (objE == null) {
+						System.out.println("Map " + idx + " is missing in (2): " + toString(objS));
+					} else {
+						if (//!objS.getMapIndex().decodeType(objS.getTypes()[0]).tag.equals(OSMAND_CHANGE_TAG) &&
+								!objE.compareBinary(objS)) {
+							System.out.println("Map " + idx + " is not equal: " + toString(objS) + " != " + toString(objE));
+						}
+						endData.remove(idx);
+					}
+				} else {
+					if (objE == null) {
+						// Object with this id is not present in the second obf
+						BinaryMapDataObject obj = new BinaryMapDataObject(idx, objS.getCoordinates(), null,
+								objS.getObjectType(), objS.isArea(), new int[] { deleteId }, null);
+						endData.put(idx, obj);
+					} else if (objE.compareBinary(objS)) {
+						endData.remove(idx);
+					}
+				}
+			}
+			if(print) {
+				for (BinaryMapDataObject e : endData.valueCollection()) {
+					System.out.println("Map " + e.getId() + " is missing in (1): " + toString(e));
 				}
 			}
 		}
+
 	}
 	
-	private void compareRouteData(ObfFileInMemory fStart, ObfFileInMemory fEnd) {
+	private String toString(BinaryMapDataObject objS) {
+		StringBuilder s = new StringBuilder();
+		BinaryInspector.printMapDetails(objS, s, false);
+		return s.toString();
+	}
+
+	private void compareRouteData(ObfFileInMemory fStart, ObfFileInMemory fEnd, boolean print) {
 		RouteRegion ri = fEnd.getRouteIndex();
 		int deleteId = ri.searchRouteEncodingRule(OSMAND_CHANGE_TAG, OSMAND_CHANGE_VALUE);
 		if (deleteId == -1) {
 			deleteId = ri.routeEncodingRules.size();
+			if(deleteId == 0) {
+				deleteId = 1;
+			}
 			ri.initRouteEncodingRule(deleteId, OSMAND_CHANGE_TAG, OSMAND_CHANGE_VALUE);
 		}
 
@@ -125,16 +159,32 @@ public class ObfDiffGenerator {
 		for (Long idx : startData.keys()) {
 			RouteDataObject objE = endData.get(idx);
 			RouteDataObject objS = startData.get(idx);
-			if (objE == null) {
-				// Object with this id is not present in the second obf
-				RouteDataObject rdo = new RouteDataObject(ri);
-				rdo.id = objS.id;
-				rdo.pointsX = objS.pointsX; 
-				rdo.pointsY = objS.pointsY;
-				rdo.types = new int[] {deleteId };
-				endData.put(idx, rdo);
-			} else if (objE.compareRoute(objS)) {
-				endData.remove(idx);
+			if (print) {
+				if (objE == null) {
+					System.out.println("Route " + idx + " is missing in (2): " + objS);
+				} else {
+					if (!objE.compareRoute(objS)) {
+						System.out.println("Route " + idx + " is not equal: " + objS + " != " + objE);
+					}
+					endData.remove(idx);
+				}
+			} else {
+				if (objE == null) {
+					// Object with this id is not present in the second obf
+					RouteDataObject rdo = new RouteDataObject(ri);
+					rdo.id = objS.id;
+					rdo.pointsX = objS.pointsX;
+					rdo.pointsY = objS.pointsY;
+					rdo.types = new int[] { deleteId };
+					endData.put(idx, rdo);
+				} else if (objE.compareRoute(objS)) {
+					endData.remove(idx);
+				}
+			}
+		}
+		if(print) {
+			for(RouteDataObject e: endData.valueCollection()) {
+				System.out.println("Route " + e.getId() + " is missing in (1): " + e);
 			}
 		}
 	}
